requirements.txt
# AI/ML Dependencies
langchain==0.1.0
openai==1.3.0
pinecone-client==2.2.4
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2

# AWS Dependencies
boto3==1.34.0
awscli==1.29.0
aws-sam-cli==1.101.0

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
tqdm==4.66.1
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Monitoring
prometheus-client==0.19.0
grafana-api==2.1.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.0.1
pytesseract==0.3.10
Pillow==10.1.0 